+++ postroj/api.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 # (c) 2022 Andreas Motl <andreas.motl@cicerops.de>
 import logging
+from pathlib import Path
 from typing import List
 
 from postroj.image import ImageProvider
@@ -29,3 +30,19 @@ def pull_multiple_images(names: List[str]):
         except Exception as ex:
             logger.error(f"Failed pulling image {name}. Reason: {ex}")
     return providers
+
+
+def pull_curated_image(image: str) -> Path:
+    """
+    Pull curated operating system image, like `debian-bullseye`, `ubuntu-jammy`,
+    `fedora-37`, or `opensuse-tumbleweed`. See also `postroj list-images`.
+
+    Return path to image directory.
+    """
+
+    # Figure out the image from the list of available ones.
+    dist = find_distribution(image)
+
+    # Acquire rootfs filesystem image.
+    ip = ImageProvider(distribution=dist)
+    return ip.image